indent blocks for numbered lists
Some of the numbered list elements are multiple paragraphs. Some of
those paragraphs are not indented far enough to line up with the numbers
of each section. This makes it difficult for readers to quickly
understand where they are mentally in the outline - this section
belongs to *which* numbered list element?

Indent each section so that all the paragraphs for a numbered list
element align under that number.